+++ packages/dart/lib/sdk_interop.dart
@@ -1,4 +1,5 @@
 @JS()
+/// experimental feature to build dalk.io SDK to JS library
 library dalk_sdk;
 
 import 'dart:html';
@@ -79,8 +80,8 @@ void main() {
   setProperty(
       window,
       'dalkSendMessage',
-      allowInterop((String conversationId, String message) => Promise<void>(allowInterop((resolve, reject) {
-            conversations.firstWhere((conv) => conv.id == conversationId, orElse: () => null)?.sendMessage(message)?.then(resolve)?.catchError(reject);
+      allowInterop((String conversationId, String message, Map<String, dynamic> metadata) => Promise<void>(allowInterop((resolve, reject) {
+            conversations.firstWhere((conv) => conv.id == conversationId, orElse: () => null)?.sendMessage(message: message, metadata: metadata)?.then(resolve)?.catchError(reject);
           }))));
 
   setProperty(