+++ packages/dart/lib/src/sdk_base.freezed.dart
@@ -12,15 +12,24 @@ T _$identity<T>(T value) => value;
 class _$_MessageTearOff {
   const _$_MessageTearOff();
 
-  __Message _(String id, String senderId, String text, DateTime createdAt,
-      MessageStatus status, List<UserMessageStatus> statusDetails) {
+  __Message _(
+      String id,
+      String senderId,
+      @nullable String text,
+      DateTime createdAt,
+      @nullable DateTime updatedAt,
+      MessageStatus status,
+      List<UserMessageStatus> statusDetails,
+      Map<String, dynamic> metadata) {
     return __Message(
       id,
       senderId,
       text,
       createdAt,
+      updatedAt,
       status,
       statusDetails,
+      metadata,
     );
   }
 }
@@ -31,10 +40,14 @@ const _$Message = _$_MessageTearOff();
 mixin _$_Message {
   String get id;
   String get senderId;
+  @nullable
   String get text;
   DateTime get createdAt;
+  @nullable
+  DateTime get updatedAt;
   MessageStatus get status;
   List<UserMessageStatus> get statusDetails;
+  Map<String, dynamic> get metadata;
 
   _$MessageCopyWith<_Message> get copyWith;
 }
@@ -45,10 +58,12 @@ abstract class _$MessageCopyWith<$Res> {
   $Res call(
       {String id,
       String senderId,
-      String text,
+      @nullable String text,
       DateTime createdAt,
+      @nullable DateTime updatedAt,
       MessageStatus status,
-      List<UserMessageStatus> statusDetails});
+      List<UserMessageStatus> statusDetails,
+      Map<String, dynamic> metadata});
 }
 
 class __$MessageCopyWithImpl<$Res> implements _$MessageCopyWith<$Res> {
@@ -64,8 +79,10 @@ class __$MessageCopyWithImpl<$Res> implements _$MessageCopyWith<$Res> {
     Object senderId = freezed,
     Object text = freezed,
     Object createdAt = freezed,
+    Object updatedAt = freezed,
     Object status = freezed,
     Object statusDetails = freezed,
+    Object metadata = freezed,
   }) {
     return _then(_value.copyWith(
       id: id == freezed ? _value.id : id as String,
@@ -73,10 +90,15 @@ class __$MessageCopyWithImpl<$Res> implements _$MessageCopyWith<$Res> {
       text: text == freezed ? _value.text : text as String,
       createdAt:
           createdAt == freezed ? _value.createdAt : createdAt as DateTime,
+      updatedAt:
+          updatedAt == freezed ? _value.updatedAt : updatedAt as DateTime,
       status: status == freezed ? _value.status : status as MessageStatus,
       statusDetails: statusDetails == freezed
           ? _value.statusDetails
           : statusDetails as List<UserMessageStatus>,
+      metadata: metadata == freezed
+          ? _value.metadata
+          : metadata as Map<String, dynamic>,
     ));
   }
 }
@@ -88,10 +110,12 @@ abstract class _$_MessageCopyWith<$Res> implements _$MessageCopyWith<$Res> {
   $Res call(
       {String id,
       String senderId,
-      String text,
+      @nullable String text,
       DateTime createdAt,
+      @nullable DateTime updatedAt,
       MessageStatus status,
-      List<UserMessageStatus> statusDetails});
+      List<UserMessageStatus> statusDetails,
+      Map<String, dynamic> metadata});
 }
 
 class __$_MessageCopyWithImpl<$Res> extends __$MessageCopyWithImpl<$Res>
@@ -108,48 +132,58 @@ class __$_MessageCopyWithImpl<$Res> extends __$MessageCopyWithImpl<$Res>
     Object senderId = freezed,
     Object text = freezed,
     Object createdAt = freezed,
+    Object updatedAt = freezed,
     Object status = freezed,
     Object statusDetails = freezed,
+    Object metadata = freezed,
   }) {
     return _then(__Message(
       id == freezed ? _value.id : id as String,
       senderId == freezed ? _value.senderId : senderId as String,
       text == freezed ? _value.text : text as String,
       createdAt == freezed ? _value.createdAt : createdAt as DateTime,
+      updatedAt == freezed ? _value.updatedAt : updatedAt as DateTime,
       status == freezed ? _value.status : status as MessageStatus,
       statusDetails == freezed
           ? _value.statusDetails
           : statusDetails as List<UserMessageStatus>,
+      metadata == freezed ? _value.metadata : metadata as Map<String, dynamic>,
     ));
   }
 }
 
 class _$__Message implements __Message {
-  const _$__Message(this.id, this.senderId, this.text, this.createdAt,
-      this.status, this.statusDetails)
+  const _$__Message(this.id, this.senderId, @nullable this.text, this.createdAt,
+      @nullable this.updatedAt, this.status, this.statusDetails, this.metadata)
       : assert(id != null),
         assert(senderId != null),
-        assert(text != null),
         assert(createdAt != null),
         assert(status != null),
-        assert(statusDetails != null);
+        assert(statusDetails != null),
+        assert(metadata != null);
 
   @override
   final String id;
   @override
   final String senderId;
   @override
+  @nullable
   final String text;
   @override
   final DateTime createdAt;
   @override
+  @nullable
+  final DateTime updatedAt;
+  @override
   final MessageStatus status;
   @override
   final List<UserMessageStatus> statusDetails;
+  @override
+  final Map<String, dynamic> metadata;
 
   @override
   String toString() {
-    return '_Message._(id: $id, senderId: $senderId, text: $text, createdAt: $createdAt, status: $status, statusDetails: $statusDetails)';
+    return '_Message._(id: $id, senderId: $senderId, text: $text, createdAt: $createdAt, updatedAt: $updatedAt, status: $status, statusDetails: $statusDetails, metadata: $metadata)';
   }
 
   @override
@@ -166,11 +200,17 @@ class _$__Message implements __Message {
             (identical(other.createdAt, createdAt) ||
                 const DeepCollectionEquality()
                     .equals(other.createdAt, createdAt)) &&
+            (identical(other.updatedAt, updatedAt) ||
+                const DeepCollectionEquality()
+                    .equals(other.updatedAt, updatedAt)) &&
             (identical(other.status, status) ||
                 const DeepCollectionEquality().equals(other.status, status)) &&
             (identical(other.statusDetails, statusDetails) ||
                 const DeepCollectionEquality()
-                    .equals(other.statusDetails, statusDetails)));
+                    .equals(other.statusDetails, statusDetails)) &&
+            (identical(other.metadata, metadata) ||
+                const DeepCollectionEquality()
+                    .equals(other.metadata, metadata)));
   }
 
   @override
@@ -180,8 +220,10 @@ class _$__Message implements __Message {
       const DeepCollectionEquality().hash(senderId) ^
       const DeepCollectionEquality().hash(text) ^
       const DeepCollectionEquality().hash(createdAt) ^
+      const DeepCollectionEquality().hash(updatedAt) ^
       const DeepCollectionEquality().hash(status) ^
-      const DeepCollectionEquality().hash(statusDetails);
+      const DeepCollectionEquality().hash(statusDetails) ^
+      const DeepCollectionEquality().hash(metadata);
 
   @override
   _$_MessageCopyWith<__Message> get copyWith =>
@@ -192,23 +234,31 @@ abstract class __Message implements _Message {
   const factory __Message(
       String id,
       String senderId,
-      String text,
+      @nullable String text,
       DateTime createdAt,
+      @nullable DateTime updatedAt,
       MessageStatus status,
-      List<UserMessageStatus> statusDetails) = _$__Message;
+      List<UserMessageStatus> statusDetails,
+      Map<String, dynamic> metadata) = _$__Message;
 
   @override
   String get id;
   @override
   String get senderId;
   @override
+  @nullable
   String get text;
   @override
   DateTime get createdAt;
   @override
+  @nullable
+  DateTime get updatedAt;
+  @override
   MessageStatus get status;
   @override
   List<UserMessageStatus> get statusDetails;
+  @override
+  Map<String, dynamic> get metadata;
   @override
   _$_MessageCopyWith<__Message> get copyWith;
 }